+++ test/dummy/app/controllers/home_controller.rb
@@ -1,0 +1,3 @@
+class HomeController < ApplicationController
+  allow_unauthenticated_access
+end