+++ test/dummy/app/controllers/protected_controller.rb
@@ -1,0 +1,2 @@
+class ProtectedController < ApplicationController
+end