+++ test/dummy/config/environments/development.rb
@@ -3,8 +3,6 @@ require "active_support/core_ext/integer/time"
 Rails.application.configure do
   # Settings specified here will take precedence over those in config/application.rb.
 
-  # config.session_store :stored_session_store, key: "_dummy_session", silence: false
-
   # Make code changes take effect immediately without server restart.
   config.enable_reloading = true