+++ test/dummy/config/initializers/session_store.rb
@@ -1,0 +1,1 @@
+Rails.application.config.session_store :stored_session_store