+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Add log subscriber events (<https://github.com/tbhb/stored_session/pull/22>)
 - Extract session model behavior into concern (<https://github.com/tbhb/stored_session/pull/24>)
+- Add integration test for Rails authentication generator (<https://github.com/tbhb/stored_session/pull/26>)
 
 ## [0.2.0](https://github.com/tbhb/stored_session/releases/tag/v0.2.0)